+++ Clowns' F8/ParticlePool.cpp
@@ -27,7 +27,7 @@ ParticlePool::~ParticlePool()
 	particleArray = nullptr;
 }
 
-void ParticlePool::Generate(std::pair<int, int> pos, float startSpeed, float endSpeed, float angle, float rotSpeed, float startSize, float endSize, uint life, SDL_Rect textureRect, SDL_Color startColor, SDL_Color endColor, SDL_BlendMode blendMode, bool vortexSensitive)
+void ParticlePool::Generate(std::pair<float, float> pos, float startSpeed, float endSpeed, float angle, float rotSpeed, float startSize, float endSize, uint life, SDL_Rect textureRect, SDL_Color startColor, SDL_Color endColor, SDL_BlendMode blendMode, bool vortexSensitive)
 {
 	// Check if the pool is not full
 	assert(firstAvailable != nullptr);
